#63: Fix

+++ backend/src/main/java/com/nhom4/moviereservation/repository/GenreRepository.java
@@ -1,9 +1,10 @@
 package com.nhom4.moviereservation.repository;
 
-import com.nhom4.moviereservation.model.Genre;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.nhom4.moviereservation.model.Genre;
+
 @Repository
-public interface GenreRepository extends JpaRepository<Genre, Long> {
+public interface GenreRepository extends JpaRepository<Genre, Integer> {
 }
