+++ backend/src/main/java/com/nhom4/moviereservation/repository/MovieGenreRepository.java
@@ -1,9 +1,10 @@
 package com.nhom4.moviereservation.repository;
 
-import com.nhom4.moviereservation.model.MovieGenre;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.nhom4.moviereservation.model.MovieGenre;
+
 @Repository
-public interface MovieGenreRepository extends JpaRepository<MovieGenre, Long> {
+public interface MovieGenreRepository extends JpaRepository<MovieGenre, Integer> {
 }
