+++ backend/src/main/java/com/nhom4/moviereservation/repository/MovieRepository.java
@@ -1,9 +1,10 @@
 package com.nhom4.moviereservation.repository;
 
-import com.nhom4.moviereservation.model.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.nhom4.moviereservation.model.Movie;
+
 @Repository
-public interface MovieRepository extends JpaRepository<Movie, Long> {
+public interface MovieRepository extends JpaRepository<Movie, Integer> {
 }
