+++ backend/src/main/java/com/nhom4/moviereservation/repository/UserRepository.java
@@ -1,12 +1,14 @@
 package com.nhom4.moviereservation.repository;
 
-import com.nhom4.moviereservation.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.nhom4.moviereservation.model.User;
+
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
     Optional<User> findById(Long id);
     void deleteById(Long id);
